Add gitinfo2 support.
gitinfo2 is a LaTeX package that shows git info (e.g., branch, commit
hash, and release date) in the footer of the compiled document.  It's
quite handy to keep track of which version of a document was sent out to
whom.

gitinfo2 support can be turned on by setting the variable \GITINFO in
confpaper.cls to "yes".  In addition, the paper's working directory must
be a non-empty git repository for this to work.

Running "make" automatically installs and runs gitinfo2.  gitinfo2 is
also run as a post-checkout git hook.